+++ src/structures/Core.js
@@ -57,9 +57,9 @@ module.exports = class extends Client {
     if (
       db !== undefined &&
       db !== null &&
-      (typeof db !== 'function' || !(db instanceof DB))
+      (typeof db !== 'object' || !(db instanceof DB))
     ) {
-      const err = 'Db option must be instance of DB.';
+      const err = 'Db property must be instance of DB.';
       throw new Error(err);
     }
     if (