+++ src/util/Model.js
@@ -9,15 +9,24 @@ module.exports = class Model {
     }
     name = name.toLowerCase();
 
-    this._defaults = defaults;
-    this._name = name;
+    this.defaults = defaults;
+    this.name = name;
+    this._modelName = `${this._name
+      .slice(0, 1)
+      .toUpperCase()}${this._name.slice(1).toLowerCase()}`;
     this.collection = new Collection();
-    this._Schema = new mongoose.Schema(options);
-    this._Model = mongoose.Model(name, this._Schema);
+    this.Schema = new mongoose.Schema(options);
+    this.Model = mongoose.model(this._modelName, this.Schema, this._name);
+    this._db = this.Model.db;
+    this._toCollection();
   }
 
+  /**
+   * @returns {Collection}
+   * @async
+   */
   async getAll() {
-    const data = await this._Model.find({});
+    const data = await this.Model.find({});
     if (!data) return new Collection();
     const col = new Collection();
     for (const val of data) {
@@ -26,44 +35,107 @@ module.exports = class Model {
     return col;
   }
 
+  /**
+   * @private
+   */
   async _toCollection() {
     this.collection = await this.getAll();
   }
 
+  /**
+   * @property {function|object|*} query
+   * @property {*} value
+   * @returns {Boolean}
+   * @example model.hasOne({ id: '1' });
+   * @example model.hasOne(value => value.id === '1');
+   * @example model.hasOne('id', '1');
+   */
   hasOne(query, value) {
     return !!this.collection.find(query, value);
   }
 
+  /**
+   * @property {function|object|*} query
+   * @property {*} value
+   * @returns {*}
+   * @example model.findOne({ id: '1' });
+   * @example model.findOne(value => value.id === '1');
+   * @example model.findOne('id', '1');
+   */
   findOne(query, value) {
     const data = this.collection.find(query, value);
-    return data ? { ...this._defaults, ...data } : this._defaults;
+    return data ? { ...this.defaults, ...data } : this.defaults;
   }
 
+  /**
+   * @property {*} data
+   * @returns {*} data
+   * @example model.insertOne({ id: '1' });
+   */
   insertOne(data) {
     if (typeof data !== 'object') {
       const text = `First argument must be an object. Instead got ${typeof data}`;
       throw new TypeError(text);
     }
-    data = { ...this._defaults, data };
+    data = { ...this.defaults, ...data };
     if (!data._id) data._id = new mongoose.mongo.ObjectID();
     const col = mongoose.connection.collection(this._name);
     this.collection.set(data._id, data);
     col.insertOne(data);
+    return data;
   }
 
-  updateOne(query, dataOrValue, newData) {
+  /**
+   * @property {function|object|*} query
+   * @property {*} value
+   * @returns {*} Deleted data.
+   * @example model.deleteOne({ id: '1' });
+   */
+  deleteOne(query, value) {
     if (typeof query === 'string') {
-      if (typeof dataOrValue === 'undefined') {
+      if (typeof value === 'undefined') {
         const text = 'Value must be specified.';
         throw new Error(text);
       }
       const prop = query;
       query = {};
-      query[prop] = dataOrValue;
-      dataOrValue = newData;
+      query[prop] = value;
     }
-    if (typeof dataOrValue !== 'object') {
-      const text = `Data must be an object. Instead got ${typeof dataOrValue}`;
+    if (typeof query !== 'object' && typeof query !== 'function') {
+      const text = `First argument must be an object, function or string. Instead got ${typeof query}.`;
+      throw new TypeError(text);
+    }
+    if (!this.hasOne(query)) return null;
+    const data = this.findOne(query);
+    if (!data) return null;
+    this.collection.delete(data._id);
+    const col = mongoose.connection.collection(this._name);
+    col.deleteOne({ _id: data._id });
+    return data;
+  }
+
+  /**
+   * @property {function|object|*} query
+   * @property {*} value Value or data.
+   * @property {*} newData
+   * @returns {*} Updated data.
+   * @example model.updateOne({ id: '1' }, { id: '2' });
+   * @example model.updateOne(value => value.id === '1', { id: '2' });
+   * @example model.updateOne('id', '1', { id: '2' });
+   */
+  updateOne(query, value, newData) {
+    if (typeof query === 'string') {
+      if (typeof value === 'undefined') {
+        const text = 'Value must be specified.';
+        throw new Error(text);
+      }
+      const prop = query;
+      query = {};
+      query[prop] = value;
+      value = newData;
+    }
+    if (typeof value !== 'object') {
+      const text = `Data must be an object. Instead got ${typeof value}`;
       throw new TypeError(text);
     }
     if (typeof query !== 'object' && typeof query !== 'function') {
@@ -72,29 +144,35 @@ module.exports = class Model {
     }
     if (!this.hasOne(query)) return null;
     const data = this.findOne(query);
-    this.collection.set(data._id, {
-      ...this._defaults,
-      ...data,
-      ...dataOrValue,
-    });
-    this._Model.updateOne({ _id: data._id }, { $set: dataOrValue });
-    return true;
+    this.collection.set(data._id, value);
+    this.Model.updateOne({ _id: data._id }, { $set: value });
+    return value;
   }
 
-  upsertOne(query, dataOrValue, newData) {
-    const updated = this.updateOne(query, dataOrValue, newData);
-    if (updated) return true;
+  /**
+   * @property {function|object|*} query
+   * @property {*} value Value or data.
+   * @property {*} newData
+   * @returns {*} Updated / Inserted data.
+   * @example model.upsertOne({ id: '1' }, { id: '2' });
+   * @example model.upsertOne(value => value.id === '1', { id: '2' });
+   * @example model.upsertOne('id', '1', { id: '2' });
+   */
+  upsertOne(query, value, newData) {
+    const updated = this.updateOne(query, value, newData);
+    if (updated) return updated;
     if (typeof query === 'function') query = {};
     if (typeof query === 'string') {
-      if (typeof dataOrValue === 'undefined') {
+      if (typeof value === 'undefined') {
         const text = 'Value must be specified.';
         throw new Error(text);
       }
       const prop = query;
       query = {};
-      query[prop] = dataOrValue;
-      dataOrValue = newData;
+      query[prop] = value;
+      value = newData;
     }
-    this.insertOne({ ...query, ...dataOrValue });
+    this.insertOne({ ...this.defaults, ...query, ...value });
+    return { ...this.defaults, ...query, ...value };
   }
 };